+++ include/events/event_subscriber.h
@@ -18,6 +18,7 @@
 #include "platform/buffer_pool.h"
 #include "platform/containers.h"
 #include "transport/endpoint.h"
+#include "transport/message_rejection.h"
 
 #ifdef SOMEIP_STATIC_ALLOC
 #include "static_config.h"
@@ -156,6 +157,14 @@ public:
      * @return true if ready for event subscriptions
      */
     bool is_ready() const;
+
+    /**
+     * @brief Observe structural rejections of incoming messages.
+     *
+     * @implements REQ_TRANSPORT_026
+     */
+    void set_message_rejection_handler(
+        platform::Function<void(const transport::MessageRejectionInfo&)> handler);
 
     /**
      * @brief Get subscriber statistics